requirements.txt
fastapi
uvicorn
prometheus-fastapi-instrumentator
psycopg2-binary